+++ cmd/routerd/main.go
@@ -49,6 +49,7 @@ var (
 var errNoIPv6PrefixAvailable = errors.New("no IPv6 prefix available")
 
 func main() {
+	detachSystemdNotifySocket(os.Args[1:])
 	if err := run(os.Args[1:], os.Stdout, os.Stderr); err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)